Montada/requirements.txt
Django>=5.2.9
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0



